LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/path-objects/2d.path.rect.winding-expected.txt

2d.path.rect.winding


FAIL OffscreenCanvas test: 2d.path.rect.winding assert_equals: Red channel of the pixel at (25, 12) expected 0 but got 255
